@@ -1,7 +1,5 @@
 import { defineConfig } from "vitest/config";
-
-// @ts-ignore
-import { sharedConfig } from "vitest-config-custom";
+const { sharedConfig } = require("vitest-config-custom");
 
 export default defineConfig({
   ...sharedConfig,
