@@ -1,13 +1,11 @@
 // Copyright Tharsis Labs Ltd.(Evmos)
 // SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/apps/blob/main/LICENSE)
-const { version } = require('./package.json');
+const { version } = require("./package.json");
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   images: {
-    domains: [
-      'storage.evmos.org',
-    ],
+    domains: ["storage.evmos.org"],
   },
   reactStrictMode: true,
   basePath: "/governance",
@@ -18,6 +16,9 @@ const nextConfig = {
   },
   publicRuntimeConfig: {
     version,
+  },
+  eslint: {
+    ignoreDuringBuilds: true,
   },
 };
 
