@@ -4,7 +4,11 @@
 import "./globals.css";
 import "evmos-wallet/styles.css";
 import type { AppProps } from "next/app";
+import { appWithTranslation } from "next-i18next";
 
-export default function App({ Component, pageProps }: AppProps) {
+export default appWithTranslation(function App({
+  Component,
+  pageProps,
+}: AppProps) {
   return <Component {...pageProps} />;
-}
+});
