@@ -1,10 +1,6 @@
 import { defineConfig } from "vitest/config";
+const { sharedConfig } = require("vitest-config-custom");
 
-// eslint-disable-next-line
-// @ts-ignore
-import { sharedConfig } from "vitest-config-custom";
-
-// eslint-disable-next-line
 export default defineConfig({
   ...sharedConfig,
 });
